requirements.txt
fastapi
uvicorn[standard]
twilio
openai
python-multipart
websockets
python-dotenv
pyngrok
